[1.0.0] User registration

+++ tests/src/test/java/org/jrestful/tests/web/controllers/rest/TestHelper.java
@@ -2,6 +2,8 @@ package org.jrestful.tests.web.controllers.rest;
 
 import org.jrestful.tests.components.article.ArticleService;
 import org.jrestful.tests.components.user.UserService;
+import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +20,8 @@ import org.springframework.web.context.WebApplicationContext;
 @WebAppConfiguration
 public abstract class TestHelper {
 
+  private static boolean initialized = false;
+
   protected MockMvc mockMvc;
 
   @Autowired
@@ -33,10 +37,22 @@ public abstract class TestHelper {
   protected UserService userService;
 
   @Before
-  public void setUp() {
-    mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).addFilters(springSecurityFilterChain).build();
-    articleService.deleteAll();
-    userService.deleteAll();
+  public void before() {
+    if (!initialized) {
+      mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).addFilters(springSecurityFilterChain).build();
+      articleService.deleteAll();
+      userService.deleteAll();
+      initialized = true;
+    }
+  }
+
+  @After
+  public void after() {
+  }
+
+  @AfterClass
+  public static void tearDown() {
+    initialized = false;
   }
 
 }
